move Metal sweepers to internal and undo other changes
