+++ equinix/resource_metal_organization.go
@@ -1,8 +1,6 @@
 package equinix
 
 import (
-	"fmt"
-	"log"
 	"regexp"
 
 	equinix_errors "github.com/equinix/terraform-provider-equinix/internal/errors"
@@ -12,7 +10,6 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
-	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/packethost/packngo"
 )
 
@@ -267,39 +264,4 @@ func expandMetalOrganizationAddress(address []interface{}) packngo.Address {
 	}
 
 	return transformed
-}
-
-func addMetalOrganizationSweeper() {
-	resource.AddTestSweepers("equinix_metal_organization", &resource.Sweeper{
-		Name:         "equinix_metal_organization",
-		Dependencies: []string{"equinix_metal_project"},
-		F:            testSweepOrganizations,
-	})
-}
-
-func testSweepOrganizations(region string) error {
-	log.Printf("[DEBUG] Sweeping organizations")
-	config, err := sharedConfigForRegion(region)
-	if err != nil {
-		return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting configuration for sweeping organizations: %s", err)
-	}
-	metal := config.NewMetalClient()
-	os, _, err := metal.Organizations.List(nil)
-	if err != nil {
-		return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting org list for sweeping organizations: %s", err)
-	}
-	oids := []string{}
-	for _, o := range os {
-		if isSweepableTestResource(o.Name) {
-			oids = append(oids, o.ID)
-		}
-	}
-	for _, oid := range oids {
-		log.Printf("Removing organization %s", oid)
-		_, err := metal.Organizations.Delete(oid)
-		if err != nil {
-			return fmt.Errorf("Error deleting organization %s", err)
-		}
-	}
-	return nil
 }
