+++ equinix/resource_metal_vlan.go
@@ -2,8 +2,6 @@ package equinix
 
 import (
 	"errors"
-	"fmt"
-	"log"
 	"path"
 
 	"github.com/equinix/terraform-provider-equinix/internal/converters"
@@ -13,7 +11,6 @@ import (
 	"github.com/equinix/terraform-provider-equinix/internal/config"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/packethost/packngo"
 )
 
@@ -176,53 +173,4 @@ func resourceMetalVlanDelete(d *schema.ResourceData, meta interface{}) error {
 	// TODO(displague) do we need to unassign gateway connections before delete?
 
 	return equinix_errors.FriendlyError(equinix_errors.IgnoreResponseErrors(equinix_errors.HttpForbidden, equinix_errors.HttpNotFound)(client.ProjectVirtualNetworks.Delete(id)))
-}
-
-func addMetalVlanSweeper() {
-	resource.AddTestSweepers("equinix_metal_vlan", &resource.Sweeper{
-		Name:         "equinix_metal_vlan",
-		Dependencies: []string{"equinix_metal_virtual_circuit", "equinix_metal_vrf", "equinix_metal_device"},
-		F:            testSweepVlans,
-	})
-}
-
-func testSweepVlans(region string) error {
-	log.Printf("[DEBUG] Sweeping vlans")
-	config, err := sharedConfigForRegion(region)
-	if err != nil {
-		return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting configuration for sweeping vlans: %s", err)
-	}
-	metal := config.NewMetalClient()
-	ps, _, err := metal.Projects.List(nil)
-	if err != nil {
-		return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting project list for sweeping vlans: %s", err)
-	}
-	pids := []string{}
-	for _, p := range ps {
-		if isSweepableTestResource(p.Name) {
-			pids = append(pids, p.ID)
-		}
-	}
-	dids := []string{}
-	for _, pid := range pids {
-		ds, _, err := metal.ProjectVirtualNetworks.List(pid, nil)
-		if err != nil {
-			log.Printf("Error listing vlans to sweep: %s", err)
-			continue
-		}
-		for _, d := range ds.VirtualNetworks {
-			if isSweepableTestResource(d.Description) {
-				dids = append(dids, d.ID)
-			}
-		}
-	}
-
-	for _, did := range dids {
-		log.Printf("Removing vlan %s", did)
-		_, err := metal.ProjectVirtualNetworks.Delete(did)
-		if err != nil {
-			return fmt.Errorf("Error deleting vlan %s", err)
-		}
-	}
-	return nil
 }
