+++ internal/sweep/sweep_test.go
@@ -3,17 +3,18 @@ package sweep_test
 import (
 	"testing"
 
-	"github.com/equinix/terraform-provider-equinix/equinix"
+	"github.com/equinix/terraform-provider-equinix/internal/resources/metal/device"
+	"github.com/equinix/terraform-provider-equinix/internal/resources/metal/organization"
 	"github.com/equinix/terraform-provider-equinix/internal/resources/metal/project"
 	"github.com/equinix/terraform-provider-equinix/internal/resources/metal/ssh_key"
+	"github.com/equinix/terraform-provider-equinix/internal/resources/metal/user_api_key"
+	"github.com/equinix/terraform-provider-equinix/internal/resources/metal/virtual_circuit"
+	"github.com/equinix/terraform-provider-equinix/internal/resources/metal/vlan"
 	"github.com/equinix/terraform-provider-equinix/internal/resources/metal/vrf"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 )
 
 func TestMain(m *testing.M) {
-	// Register legacy sweepers for resources balled up under equinix/
-	equinix.AddMetalTestSweepers()
-
 	// Register sweepers for individual resource packages
 	addTestSweepers()
 
@@ -21,7 +22,12 @@ func TestMain(m *testing.M) {
 }
 
 func addTestSweepers() {
-	ssh_key.AddTestSweeper()
+	device.AddTestSweeper()
+	organization.AddTestSweeper()
 	project.AddTestSweeper()
+	ssh_key.AddTestSweeper()
+	user_api_key.AddTestSweeper()
+	virtual_circuit.AddTestSweeper()
+	vlan.AddTestSweeper()
 	vrf.AddTestSweeper()
 }
